Move utility functions to tools
